fix: Correct signedness mismatch in TetgenAdapter.cpp
Fixed a warning in TetgenAdapter.cpp where 'size_t' was being compared to 'int'. Replaced 'size_t' with 'int' to ensure the types being compared are of the same signedness. This change resolves the compiler warning and prevents potential issues related to signedness mismatches.

+++ src/polyhedralGravity/input/TetgenAdapter.cpp
@@ -115,7 +115,7 @@ namespace polyhedralGravity {
                             "Converting tetgen's trifaces to C++ Polyhedron");
         _faces.clear();
         _faces.reserve(_tetgenio.numberoftrifaces);
-        for (size_t i = 0; i < _tetgenio.numberoftrifaces * 3; i += 3) {
+        for (int i = 0; i < _tetgenio.numberoftrifaces * 3; i += 3) {
             _faces.push_back({static_cast<size_t>(_tetgenio.trifacelist[i]),
                               static_cast<size_t>(_tetgenio.trifacelist[i + 1]),
                               static_cast<size_t>(_tetgenio.trifacelist[i + 2])});
